frontend/requirements.txt
# GenArt Studio - Frontend Dependencies
streamlit>=1.28.0
requests>=2.31.0
pillow>=10.0.0